Update version 1.1
